refactor : transaction code refactoring
