+++ src/main/java/com/example/transaction/entity/User.java
@@ -1,0 +1,26 @@
+package com.example.transaction.entity;
+
+public class User {
+
+    private final Long id;
+    private final String name;
+    private final Integer age;
+
+    public User(Long id, String name, Integer age) {
+        this.id = id;
+        this.name = name;
+        this.age = age;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public Integer getAge() {
+        return age;
+    }
+}
